fix(ssm,quadrature): preserve input dtype under x64 (gh-219)
Two independent float64 promotions made float32 unusable end to end when
jax_enable_x64 is on. Pure float32 (x64 off) was unaffected, so this bit
mixed-precision users rather than the default path.

kalman_filter's predict-only branch returned a bare jnp.array(0.0) — a
float64 scalar under x64 — while the update branch's accumulator followed
the float32 inputs, so lax.cond rejected the branches outright. Carry the
dtype through in _skip_update and init_carry, as nonlinear_kalman_filter
already does.

The quadrature rules built their points and weights from Python and NumPy
scalars, landing in float64 regardless of the input dtype and promoting
every downstream moment. Build them in mean.dtype throughout:

- sigma_points, cubature_points, fifth_order_cubature_points: weights
- fifth_order_cubature_points: also the shell points, whose np.sqrt radii
  were strongly-typed np.float64 and promoted what they scaled
- gauss_hermite_points: takes an explicit dtype, since it has no mean to
  key off; GaussHermiteIntegrator passes the input dtype
- MonteCarloIntegrator, AssumedDensityFilter: draws, weights, and jitter

uv.lock picks up the 0.0.19 -> 0.0.21 version sync it had drifted from.

+++ src/gaussx/_quadrature/_gauss_hermite.py
@@ -67,8 +67,9 @@ class GaussHermiteIntegrator(AbstractIntegrator):
         mu = state.mean
         N = mu.shape[0]
 
-        # GH points in standard normal space
-        z, w = gauss_hermite_points(self.order, N)
+        # GH points in standard normal space, in the input dtype so a
+        # float32 belief is not promoted under x64.
+        z, w = gauss_hermite_points(self.order, N, dtype=mu.dtype)
 
         # Transform to input space: xᵢ = μ + S zᵢ
         S = sqrt(state.cov).as_matrix()
